KUDU-987 (part 2): don't acquire tablet components until applying, don't wait on non-applying transactions
This changes the flow of write transactions such that we no longer take
our copy of Tablet::components_ during the PREPARE phase. Instead,
we defer it until the beginning of the APPLY phase. Once we are in the
APPLY phase, we know that (a) the transaction will commit, and (b) the
transaction will finish within a bounded amount of time, limited only
by the speed at which the actual operations can be processed.

We also change Tablet::Flush and the pre-metadata-flush callback such that we
no longer wait for _all_ in-flight writes, but rather just those which are
already in the APPLY state in MvccManager. This is safe, because what we really
cared about in those cases was catching all writes which saw a previous version
of Tablet::components_. Since we capture Tablet::components_ only after
entering the APPLY state, we know that all other in-flights (not yet in APPLY)
will see the new 'components_' value.

This avoids the situation described in KUDU-987 where a write transaction is
replicated, acquires the components, and then a majority of the replicas
goes down. In that case, the operation would sit in REPLICATED-PREPARED state
indefinitely, causing the WaitForAllInFlights() calls in Tablet::Flush or
the pre-metadata-flush hook to wait forever. This "wait-forever" behavior
had two bad side effects: (1) we couldn't delete the tablet, because Shutdown
would be blocked by the running flush operation, and (2) the maintenance
manager thread would be blocked indefinitely, causing all other tablets
on the TS to slowly use up all of the available memory.

The new unit test is contributed by Mike. Here's what he has to say:

  Added a test that previously failed, but now passes with this series of
  changes. The failure was caused by an in-flight MRS flush that would
  have blocked forever. Attempting to shut down the MaintenanceManager
  while this flush was outstanding would cause it to block on shutdown
  waiting for its worker thread to complete the flush op. This in turn
  would cause TabletPeer::Shutdown() to hang indefinitely, which caused
  DeleteTablet() to do the same.

Because this touches a particularly delicate part of the code, I looped
mt-tablet-test 1000 times in a fastdebug build. All passed. I also looped
linked-list-test a few hundred times with no failures. Previous buggy
versions of the "flush barrier" code path would cause linked_list-test
to fail. Lastly, I looped the newly added unit test 500 times with no
failures.

Before commit, I will also test this on bolt80 using a workload which
triggered this reliably before: run tpch_real_world into a table with
240 tablets. While running the write workload, delete the table. After
the delete completes, I expect to see all of the tablets successfully
deleted, and all tablet servers should report 0 transactions in flight.

Change-Id: I5707d6395222fe2bbe9546759f76b678898209bb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7474
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>