mdl: add metrics for histogram total count and sum
Since CM can't support tracking raw histograms, we instead will just track
the total count and sum for now. These make reasonable counters for looking
at the arrival rate of events and the mean value.

This also adds the 'label' field to the JSON dump for histograms, which was
inadvertently omitted.

Change-Id: I94c6779085caabfa96195583446a49898e3808ba
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7518
Reviewed-by: Christian Leroy <cleroy@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>