full_stack-insert-scan-test: increase timeouts
Despite spending much time searching for a deadlock that would cause the
kudu-benchmarks job to fail, I couldn't find a smoking gun. But we still
want the job to be resilient, so let's raise the various timeouts associated
with the test.

If there truly is a deadlock lurking somewhere, the job will fail and we can
revisit the issue.

Change-Id: Ia3df14a1d13f178cec35b358bf86d5ebbf9abca0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7519
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins