ts-cli: return error codes instead of crashing
I intend to use ts-cli in the Kudu package init scripts to check if a server
has really started. To that end, it needs to gracefully fail instead of
CHECKing; the latter yields an ugly message in the shell that can't be
stifled.

Also, I changed --tserver_address to --server_address because ts-cli can
actually be used on masters too.

Change-Id: Ibc804dfd5a1c85a95c0c1d7603200a5647dd3309
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7522
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>