Add NVML to thirdparty
This is based on Sarah Jelinek's work. A few modifications:

- uploaded nvml 0.2 to our thirdparty-libs S3 bucket
- added code to build-thirdparty to build and install libvmem into
  thirdparty/installed/
- changed cmake_modules/FindVmem.cmake to look in thirdparty/installed
  of /usr/lib

The only subtlety to note here is that we are building only libvmem.  It looks
like Sarah's patch only uses this interface for now, and other parts of the
vmem library depend on libuuid, which we don't have yet in thirdparty. Rather
than pull in the thirdparty dependency, I decided to just build the library
that we need.

Change-Id: I294a5d1c561f296be2cf012c784f7104cc744120
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7529
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins