Workaround glibc bug in which timer_create() fails
This works around a bug in glibc[1] in which timer_create() will
fail with EAGAIN if there is a lot of static thread-local storage
in use in the program. The timer_create() implementation tries to
create a thread with a hard-coded 16KB stack, but if there is a lot
of TLS, then that's not large enough. Apparently the TLS is allocated
from the thread's stack.

The fix is gross: we override pthread_stack_setattr to enforce at least
a 32KB stack. This is enough to fix the failing tests when linking
against libvmem.so, which fails without it.

[1] https://sourceware.org/bugzilla/show_bug.cgi?id=13088

Change-Id: Icb35654867891c8788ecfd80006a82ce788190cd
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7539
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins