Add KuduSchema::GetPrimaryKeyColumnIndexes() to the public API
This is a feature requested by Binglin. A previous patch removed the
public num_key_columns() API, since we didn't want to encode an
assumption that the key columns are always the first N columns.
The new API instead returns a vector of column indexes. That vector
today happens to always be (0, 1, ..., N) but in the future the API
will not be broken when we support key columns not listed first.

Change-Id: I0c1176dae3cf8b982ac055b06716d3958eb16021
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7541
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins