KUDU-1035. Fix crasher in tablet bootstrap due to out of order timestamps
Tablet bootstrap assumes that it can move the offline safe time any time
a transaction commits, however it considers consensus-only operations
such as NO_OP and CHANGE_CONFIG_OP to be included in this as well. This
is no longer a safe assumption, because we do not assign the timestamps
for these operations on the prepare thread like we do for the tablet
transactions.

The fix is to exclude consensus-only operations from tablet bootstrap's
safe time advancement logic.

The included test crashes without the included patch to
tablet_bootstrap.cc

Change-Id: Ic0e6da041a33cae96e160fb4c682794b3f81ac1d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7545
Tested-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>