Fully remove old log directory at log recovery time
Cherry-picking files to delete is tedious and error-prone. For example,
we are not currently deleting log indexes. Removing the whole directory
at recovery time prevents us from having to remember to remove all the
random stuff that might be in there.

Added a unit test for log replay recovery, also addressing KUDU-34.

Change-Id: I7dab417cc78be4e4eb350424c94740b48eb9dbdc
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7549
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins