build: fix build when using ld, not gold
Since commit 08c1ac3, linking kudu-tserver breaks with the following error:

  /usr/bin/ld: thirdparty/installed/lib/libLLVMSupport.a(DynamicLibrary.cpp.o): undefined reference to symbol 'dlclose@@GLIBC_2.2.5'
  /usr/lib/x86_64-linux-gnu/libdl.so: error adding symbols: DSO missing from command line
  collect2: error: ld returned 1 exit status
  make[4]: *** [build/release/kudu-tserver] Error 1

The addition of dl as a first-class target caused cmake to place it earlier
in the kudu-tserver link line:

  @ -56,6 +56,7 @@
   ../../../build/release/libserver_common.a
   ../../../build/release/libtablet_proto.a
   ../../../build/release/libcodegen.a
  +/usr/lib/x86_64-linux-gnu/libdl.so
   -lLLVMIRReader
   -lLLVMInstrumentation
   -lLLVMipo
  @@ -89,7 +90,6 @@
   -lLLVMMC
   -lLLVMObject
   -lLLVMSupport
  --ldl
   -ltinfo
   -lpthread
   -lz

It's not clear to me why this happened. It has something to do with how
FindCyrusSASL.cmake adds dl as a dependency of cyrus_sasl, but beyond that I
can't figure it out.

That said, the original impetus for having dl be a first-class target is
gone, since its usage is restricted to non-"exported" targets. So we can
safely remove that code. If dl is ever needed by an exported target, we'll
have to revisit this.

Change-Id: Ib5e3261a2b7be2a40a644465b09f8aea6ec904c3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7552
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>