KUDU-699 (part 1). Peers shouldn't send more requests once Close() is pending
This changes the behavior of Peer::Close() so that, if there is a pending
request, we close the peer as soon as it responds. Previously, if the
response came and there was still more data to send, the peer would trigger
another request, causing Close() to block for arbitrarily long.

This is only 'part 1' since, ideally, we would be able to Close a peer even
while a request was in-flight. Otherwise, leader step-down can be as slow
as RPC timeout, which might cause flapping.

Change-Id: I58e30cd94055fd072afe587556d32616ae53717c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7562
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>