external_mini_cluster_fs_inspector: disregard tmp files when listing directories
Of the last 307 failures of delete_table-test, 10 of them had failures that
look like:

Bad status: Timed out: Tablet 46e6958493af4182af1e765ed5e024d8.tmp.XBP9Ua not RUNNING after 30.006s: UNKNOWN

This is because we rely on directory listing to determine the tablets that
exist. This means that, if we list the tablet metadata directory while the
metadata is in the process of being written, the tests can mistake the tmp file
name for a tablet ID.

This patch simply makes the listing ignore files with a ".tmp.*" suffix.

Change-Id: I9fed6aea25d90f13c4a20ee8a7c4138bded62179
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7565
Tested-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>