KUDU-821 (part 2): don't get stuck in a tight loop sending RPCs to deleted tablet
This fixes an issue where, if a tablet was deleted, and remote bootstrap was
disabled, the consensus queue would still mark the peer as needing a request.
This would cause a tight loop of the leader sending a request to the peer,
seeing that it was deleted, and immediately sending another.

The fix changes the handling of the remote_bootstrap_enabled flag such that
it's checked right before sending the bootstrap request. At that point, we
consider the flag being set to be an "error", which prevents retries until
the next heartbeat.

The test verifies that the deleted TS isn't hammered with UpdateConsensus calls,
and that the leader TS isn't spewing log messages. Without the bug fix in place,
both assertions fail (seeing about 35 log messages and RPCs per second in
a debug build, probably worse in a release build)

In order to make the test more reliable, I also made the metrics-fetching
utility code only fetch metrics that match the requested metric name.
This reduces the parsing cost, especially on instrumented builds like
TSAN, where fetching all of the metrics apparently takes a reasonable
long time.

Change-Id: I3cddb8d00e667c57a6bf82995b36507d154e2cce
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7582
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>