ts_tablet_manager: open all metadata before starting bootstrap
This changes the tablet manager to open all of the metadata blocks before
it begins actually running the bootstrap process for any tablet.

I noticed on some hosts which are holding a fair number of tablets that
the TSTabletManager was taking a very long time in Init() because
of slowness while loading the metadata blocks. The Init() was especially
slow because it was competing for disk resources with a bunch of
bootstrapping tablets.

No test since this just rejiggers some code already covered by existing
tests.

Change-Id: I22896104c80b5c7a04a80f62eddf053b9577a4ab
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7578
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>