KUDU-821 (part 1). If the previous request to a Peer had an error, don't send a new request for each write
Previously, if a follower was down, every write to the leader would call
Peer::SignalRequest(), and we'd see that no RPC was already in-flight.
This would cause us to start a new attempt to contact the follower.

The end result was that we'd end up sending hundreds or thousands of RPCs
per second to downed followers. On the YCSB cluster, this filled up the
log partition and caused several machines to crash.

This patch fixes the issue by disallowing the 'expedited' request to the
follower if the previous request had an error.

The test sets up a mocked peer which responds with an error, and then
calls SignalRequest() on it as if writes are being sent. We assert that
we get only a couple calls to UpdateConsensus().

Change-Id: I6cfe4d8dbb20b717129bb3b0aaae41a5efc1ebb6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7581
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>