Add metrics to count GLog messages at each level
This metric is generally useful to see if a particular server is spewing log
messages faster than others. In addition, it's also useful for testing
purposes: I'm currently attempting to write a regression test for a case where
we have log spew, and this will make it easy to assert on the rate of
log messages.

Change-Id: I5cd8fee0a29cfc7bec37f7c03110a73e7d664390
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7580
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>