KUDU-1049. Add various metrics for scanners
This adds a bunch of new metrics that I thought would be useful when
I was starting to build chart dashboards for CM:

- rows/cells/bytes returned to the client (post-MVCC, post-predicate, etc)
- rows/cells/bytes actually scanned from disk (pre-all-of-the-above)

These are tracked on a per-tablet basis. I plumbed them through the
IteratorStats class which already existed, which means that it was also easy to
expose the new information on a per-column basis on the web UI. While I was
there, I also improved the UI to display the column names instead of the column
indexes which were there before (and fairly useless).

Here's a screenshot of the new iterator info on the web UI:
https://s3.amazonaws.com/uploads.hipchat.com/108018/844762/YH4N2mW3IklEdGx/iterator-stats.png

Normally, I'd write a unit test for this, but CM is code-freezing soon, and I'd
like to get these in ASAP so we can build a dashboard that uses them. I'll
follow up with a unit test before marking the JIRA as closed. I did a manual
test by loading ~4M rows into a TS and then scanning the tablet using the
'ts-cli dump_tablet' tool. The results I got were:

{ name: "scanner_rows_returned", value: 4250000 },
{ name: "scanner_cells_returned", value: 68000000 },
{ name: "scanner_bytes_returned", value: 1065825414 },
{ name: "scanner_rows_scanned", value: 4250000 },
{ name: "scanner_cells_scanned_from_disk", value: 68000000 },
{ name: "scanner_bytes_scanned_from_disk", value: 616033265 },

This matches up with the expected values:
- I had 4.25M rows in the tablet
- The table had 16 columns, so I expected 68M cells returned
- 250 bytes per row seems reasonable (including overhead like null bitmaps,
  string Slices, etc)
- no predicates, so I expected rows_scanned and cells_scanned to match
  rows_returned and cells_returned
- the tablet is reporting 587MB as "on-disk size", so the amount of
  data scanned from disk matches up.

Change-Id: Ibcbd708e98ab6e736f6af2ce14bd406c454f41b6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7596
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins