KUDU-818: (1/many) Add Partition and PartitionSchema
This commit introduces the PartitionSchema and Partition classes. These classes
will be the basis of flexible partitioning in Kudu.

The PartitionSchema class describes the partitioning and distribution layout of
a single table, analogously to how a Schema describes the logical columns of a
table. The PartitionSchema is responsible for encoding partition keys, creating
string representations of partition keys for debug printing, and creating an
initial set of partitions during table creation. The PartitionSchema of a table
is immutable, and can not be altered after creation. A follow-up commit will add
partition pruning logic to the PartitionSchema class, which will enable
efficient scans with predicates and start/stop primary keys across a non-PK
partitioned table.

The Partition class is a simple descriptor of a contiguous chunk of the
partition keyspace. Each tablet is assigned a Partition, and is responsible for
serving the rows that fall into that Partition. A Tablet's Partition does
describe what rows are under the control of that tablet, but the rows in the
tablet remain stored and sorted in Primary Key order.

Change-Id: I0e10c4c8a1e15904275d3681033a8f8b0c99d0fd
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7246
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins