test_result_server: add 2-day failure rate and sparklines
This adds a new column to the flaky test dashboard which lists the
failure rate over the previous 2 days, in addition to the existing
rate which tracks the previous 7 days. This is useful to see if a
highly flaky test has actually been fixed more recently (or if it's
getting worse).

I also added a sparkline (very small graph) which shows the trend
over the last 7 days - it's easy to see if a test has taken an
uptick or improved.

Tested locally, and now also deployed on a1228.

Change-Id: I31e0e6bbd3d2cdc5ba99e44b335cee892c4be4db
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7597
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>