KUDU-818: (2/many) initial flexible partitioning support
This commit lays the foundation for flexible partitioning in the server, C++ client,
and Java client.  Partitions and PartitionSchema have been wired in to the RPC
calls. The Java client has full support for flexibly partitioned tables; C++ client
support for creating and scanning non-simple partitioned tables will be in a
followup patch. Efficient pruning will be in a follow up commit.

This change breaks forwards and backwards compatibility with clients, but
maintains master and tablet server compatibility with on-disk data.

Change-Id: I7627113017298417fbd4dd565335bf6f2910aeca
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7247
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>