KUDU-941 (part 1): Only delete tablet superblocks on TS startup
Tablets deleted with the TABLET_DATA_DELETED flag represent members of
tables that are being permanently deleted. These types of tablets are
never allowed to be "revived" in any meaningful way so we fully delete
them. However, in order to avoid racing against a leader noticing its
replicas are gone and attempting to remotely bootstrap its followers
while the master attempts to delete the whole table, we want to leave a
marker such that remote boostrap will fail, at least for a while.

This patch causes tablets deleted with a TABLET_DATA_DELETED flag to
leave their superblock on disk until the next startup of the tablet
server. A TODO remains, pointing to KUDU-941, mentioning that we may
want to do something a little more complex in the future, such as having
a background thread do the deletion after some delay. We will repurpose
KUDU-941 to track that change after committing this one, but this takes
most of the short term pain of the flapping away.

Change-Id: I178d81118755eb9c6c8c02505361ab6c537bf41e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7604
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>