Set default number of parallel bootstrap threads to number of disks
Previously, we were using 50 parallel bootstrap threads, which is a
pretty bad idea on a machine with only a few disks. Really, it's
probably a pretty bad idea on any machine.

The new code defaults to one thread per configured data directory,
but still allows tuning in case we are running with a small number
of fast devices (eg SSD).

I also removed a comment about setting the apply_pool based on the
number of devices, since often it is CPU-bound (for RAM-resident
workloads).

Change-Id: I1d2156e6a5c9d9a9d1dc72df62572dab08ead286
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7606
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>