Add reason string to MarkDirty callbacks
Adding a reason string gives very interesting and informative logs when
--vmodule=ts_tablet_manager=2 is passed, if one happens to be trying to
understand what is happening with MarkDirty() calls and where they are
coming from. But it is probably too noisy for LOG(INFO), so this patch
is mainly useful for debugging.

Also included are a few simple fixes to ref the Consensus object pointed
to by TabletPeer before using it that I came across while writing this
patch.

Change-Id: I2641a7783152568c65c9f0cb86193e3c21c0506b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7637
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>