KUDU-1069. Prevent leader flapping in Master metadata on config change
When a config change is reported to the master, it may not include the
leader because the follower doing the reporting may not know who the
leader is yet (it may have just started up). Currently, since the config
change will often contain a higher committed opid than the
currently-stored one, the master will overwrite its tablet config
information with the one sent by the new follower, and will "forget" who
the leader of the current term is until the next time it gets a report
from someone with that config.

As of this patch, if the master is able to determine that the term
has not changed, it assumes that the leader should not have changed
either, and retains the leader's info, even if not reported by the new
follower. It also adds a sanity check to ensure that no such leader
change was reported within the same term.

To test this, I added a test that requires that the master reports back
a leader after the config change in the test completes. Before this
patch, it would fail about 50% of the time.

Change-Id: I297395cd665de24498a8e51ef63e200ab68a492b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7638
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>