KUDU-1022. Report tablet config changes to the master
Add some missing MarkDirty() calls when the config changes. This is
especially easy to reproduce using config change to remove a server.
This patch modifies RaftConsensusITest.TestMasterNotifiedOnConfigChange
to add a removal call followed by a GetTabletLocations() master RPC.
Prior to this change, the test fails because the Master is never
notified that the replica has been removed from the config.

Change-Id: I9fa5a9a59f484ff8ae35adada1145c013971c6a5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7634
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>