KUDU-1070, KUDU-1071: mark tablets as deleted when table is deleted, don't expose deleted entities
This changes DeleteTable to actually cascade the delete to the tablet
level as well, marking those tablets as deleted. The entities JSON
is updated to not list deleted tables or tablets, so that the CM
metrics integration doesn't end up having to parse arbitrarily large
data.

Additionally, I updated the GetTabletLocations to return proper errors
for deleted tablets (before, it would just ignore them).

This patch doesn't address "roll-forward" if the master crashes during a
delete table. I left a TODO and will leave the JIRA open to address that.
But, I wanted to get the partial solution committed to unblock CM testing.

Change-Id: Ia33273768ddb43f42117e5db9079fb2ae5668298
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7645
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins