KUDU-1062. Bump timeout for master->TS retries to an hour
This changes the master to retry for up to an hour when sending
tasks to a tablet. The exponential backoff is now capped at once
a minute.

Without this patch, create-table-stress-test on replicated tablets
would often fail, because some deletions would time out and thus
the tablets would never be deleted.

No new test since the create-table-stress-test patch coming after
this one exposes the bug.

Per Adar's request, I added flag tags to the advanced options inside
the catalog manager.

Change-Id: Icfc87c14ad3f8a1aea38244e7ed78724e311ddc6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7653
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>