KUDU-953. Fix a race in HybridClock which allowed clock to move backwards
This fixes an issue where HybridClock::Update() would grab the current
clock value outside of the lock, then acquire a lock, and proceed
to act based on the previously-read value. Another update or clock
read could have happened in between, such that we were acting
on stale data. This could cause the clock to move backwards, which
would wreak havoc in clock consumers.

The patch fixes the issue (a simple fix) and adds a multi-threaded
regression test. The test reproduced the issue 100% of the time.

Change-Id: I08ee7a52969d1f435594ed35cd2c12a52d73b216
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7670
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>