Make TabletService pre-allocate space for the indirect data
We're only pre-allocating space for the direct data, which causes
us to have to grow the buffer and memcpy a lot for schemas with
lots of strings.

The fix is just to pre-allocate both buffers.

Todd ran this patch on bolt80 with an Impala workload. Before the patch,
the TS was spending 33% of its time in memcpy. After the patch, 23%.

Change-Id: Ia7aacc3af0bcba2c25b0d59865ff8256b041197f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7690
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins