Fix TSAN error in raft_consensus_quorum-test
The test peer proxy implementation submitted a callback to a threadpool
while holding onto a member variable 'lock_'. The callback in some cases
would delete the proxy itself, which got reported as a bad mutex unlock
in TSAN.

This is causing the test to be about 10% flaky. I was able to repro the
TSAN failure on my laptop in a few minutes without the patch, and wasn't
able to repro it with the patch.

Change-Id: I68f6a522e9e5f816cdd9e4dfff0a4ed0775fa121
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7702
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>