KUDU-398, KUDU-1012. Scanners should allow times to be in the future, so long as they are within max error.
This changes the service side scanner implementation to allow the caller to
specify a timestamp which is in the future relative to the local clock, so long
as that future timestamp is less than now + FLAGS_max_clock_sync_error_usc.
This is often the case when using a timestamp returned by one server to perform
a scan at another -- for example in the context of running ksck with the
checksum scan option against a cluster.

The implementation now uses the clock to wait for the timestamp to be in the
past according to the local node, and then continues with the current
behavior of waiting for the MVCC manager to consider the snapshot clean.

This does not yet address adding timeouts to these 'wait' calls. It's still
possible that we can hang forever when scanning a replica that has been
abandoned.

Change-Id: Ica5aa5c75417b72bae49c722a8488683fdc56d8f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7706
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>