Remove kuduReplicaId from MDL, make "uuid" entity field more obvious
There was some confusion between me and Chris about about "kuduReplicaId"
attribute. Kudu itself has no notion of a replica ID. I had thought that we
needed to define such an attribute which CM would "fill in", but turns out I
was wrong. Chris thought we had such a notion, but turns out he was wrong.
So, I've removed this attribute from the MDL.

The "dump-entities" page was furthering the confusion by exposing a 'uuid'
field for each replica. That UUID was in fact the tablet server's UUID. So, I
renamed that to 'server_uuid' to make it more obvious.

Change-Id: Ifa04fdf5a56095e5a423011bea697afd2d1e2b69
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7719
Reviewed-by: Christian Leroy <cleroy@cloudera.com>
Tested-by: jenkins