metrics: expose gauges which are actually counters as counters
This adds a flag which can be passed when defining a "Gauge" metric that
indicates that it's actually a counter masquerading as a gauge and should be
exposed as such.

This is useful because we allow gauge metrics to be bound to a function, but
not counters. Yet, some of the functions just act as simple accessors to
separately-maintained counters. So, we should tell external metrics customers
(e.g. CM) that they are in fact counters, not gauges.

Why does it matter? It matters because CM converts counters to rates
internally, allowing for more reasonable aggregation, etc.

Change-Id: Ia4e0584aaf7e58a67ca38fe365b5c47c5a58a0af
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7730
Reviewed-by: Christian Leroy <cleroy@cloudera.com>
Tested-by: jenkins