KUDU-1057. Re-snooze after persisting "yes" vote
The election timer could use a refactor but since we are close to beta
this is more of a band-aid. We simply re-snooze the election timer after
persisting a "yes" vote to disk.

Also add a flag to control whether the snooze will be logged to avoid
seemingly duplicate log messages.

If disk latency is higher than the election timeout, this will not help,
but if it's just close to the election timeout then this may prevent
leader-election flapping.

Change-Id: I0f9d10b1239efd3064490d25e9d2b101e4e540db
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7748
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins