KUDU-1003. Add a time-based WAL retention policy
This adds a time-based WAL retention policy, defaulting to 5 minutes of log
retention. This should be long enough that a short hiccup or rolling-restart of
a server won't end up in it being dropped from a raft configuration, but better
than a static "retain X number of log segments" in that quiet tablets will end
up with only a small number of retained logs.

A new unit test verifies the behavior. I also tried it locally using
tpch_real_world. I ran with --v=2 and verified that the new log messages looked
reasonable, and that the logs were GCed after the configured retention time.

Change-Id: I48703a6f9d84d1e557ffe694bf6ebdb82585e0a4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7736
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins