KUDU-818 (3/many): support creating, scanning hash-bucketed tables in C++ client
This adds APIs to the C++ client to create a hash-bucketed table, or a table
partitioned on a range key different than the primary key.

The scanner code is also modified so that, if the table is one of these fancier
partitioning options, it doesn't attempt to do tablet pruning.

A new integration test creates tables with many different partitioning schemes,
inserts rows, and performs various scans on them to verify that the read path
works correctly.

Future work:
- we should consider changing some of our demos like TPCH and twitter to use
  bucketed tables
- we should support tablet pruning based on equality predicates on bucketed
  columns.
- we should support explicit APIs for scanning ranges of partition keys, which
  will be necessary for Impala integration.

Change-Id: I46b30ee2fa338c9d54c2c0511e34c26c85b0b61f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7765
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>