KUDU-785. Allow setting cfile block size per column
This adds APIs in Java and C++ to specify the cfile block size
on a per-column basis.

Setting the block size improves random access performance significantly.
I ran YCSB against a local tablet server with "workload b", first with
the default block size, and then again with a 4KB block size. In both
cases, I loaded the data, then waited until the maintenance manager
dashboard stopped showing flushes and compactions. I then compared
the statistics printed out after the benchmark hit steady-state:

256KB block size:

  2015-09-02 15:50:55:317 130 sec: 630079 operations; 4899.6 current ops/sec; [READ AverageLatency(us)=3315.91] [UPDATE AverageLatency(us)=2174.17]

4KB block size:

  2015-09-02 15:55:47:884 20 sec: 408341 operations; 22099.7 current ops/sec; [READ AverageLatency(us)=721.41] [UPDATE AverageLatency(us)=722.02]

This is a ~4.5x speedup for this workload. Giving users the ability
to configure this on a per-column basis will be very useful for folks
with random-access components of their workload.

Change-Id: I1c8526173e6eb79dee42924423fb654e231861ed
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7791
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins