java: fix auto-detection of build dir when path has special characters
This should fix occasional gerrit failures where the tests can't find
the Kudu binaries. The issue was that Jenkins sometimes decides to
make the workspace directory have an '@' in it. This got encoded as
'%40' in the java.net.URL object representing the path, and URL.getPath()
didn't un-escape it.

Tested locally by moving my java directory to 'java@x'. Without the
patch, it failed.

Change-Id: I5507d72d41da687d604a8f6419e11808b0bce9be
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7794
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins