Validate that leader is a voter in the active config in queue
This patch adds several safety checks to the queue and makes the leader
queue aware of which peers are in the active config. It also enforces
that peers must be a member of the active config to be registered with
the leader queue. This helps ensure correctness of the queue counts.

This was originally a refactor that almost got dropped, but the
additional checks seemed useful enough to keep.

In order to keep these checks, the order in which we reset the peer
manager and the queue in RaftConsensus had to be slightly modified. The
reasoning is in a comment there.

Also change a couple of tests that used a nonsensical majority size of 1
to test with realistic config sizes.

Change-Id: I4d48f8850dc602c6238a6d6fc72efe16c178d58c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7803
Tested-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>