KUDU-1096 (part 1): Evict abandoned followers
When followers fall too far behind the leader's log, remove them from
the raft config. Another patch is on the way for doing the same thing
with followers that are believed to be offline.

The flag to enable this, --evict_failed_followers, is left as false
for now, until we get the Master side of things (replacing nodes in
weakened configs) committed.

Change-Id: Ie2bb56b1874f74fa6523eadd28adaab9939d05e7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7805
Tested-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>