KUDU-1110. Better respect client-provided timeout on non-fault-tolerant scans
Currently, we use the Client's "default_rpc_timeout" as the timeout for
every Scan RPC, even if the scanner's timeout itself was significantly
longer. The original thinking here was that, if one server had hung,
using a shorter timeout allowed us time to fail over and try another
server.

However, this only makes sense if we actually are able to failover. If
we're already mid-tablet, then we can only failover if the scanner is
fault-tolerant. If we're opening a new tablet, we can failover any time
there is another replica available.

This modifies the code to obey the above smarter policy. I verified on
bolt80 that this fixed timeouts I was seeing when running a large
'CREATE TABLE x STORED AS PARQUET SELECT * FROM kudu_table'. This query
did enough HDFS write IO on the Parquet side that scans occasionally
went over the 5sec default_rpc_timeout on the read side.

Change-Id: I4d238238a2d9ae99a5bcea7fd5fb96225838fb30
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7834
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>