KUDU-1096 (part 2): Master should re-replicate weakened tablets
This patch adds support to the Master for an operation that initiates
a config change to add a server to a tablet config that has less than
the number of replicas specified in the Master's metadata for the table.

This patch also modifies the RetryingTSRpcTask API slightly to allow for
"failing" and retrying an RPC without actually sending it in the first
place, due to for example problems generating the request.

Change-Id: Idf5bf51d94060a45e9c1e98dbdfe797a5b6eb6ff
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7862
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>