status: add a static assertion on Code enum's size
Technically, the compiler is free to choose a different size for the private
'code' enum inside of Status based on the defined values. In practice,
compilers seem to choose 'int' sizes. Since the code enum is used within inline
functions in the header, and the Status class is part of our public ABI, it's
important to guarantee this enum doesn't change sizes.

This adds a static assertion that the code enum is given 4 bytes, so that if we
tried to build on a platform with a different enum size, we'd get a compilation
error instead of a hard-to-debug crash.

Change-Id: I1b0266df9f442dc9b46172f5071d4338eb124772
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7865
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>