KUDU-1099. Count DNS resolution failures as failed votes in leader election
Previously, we had a CHECK_OK which crashed the server if the leader failed to
resolve the host name of one its followers while calling a leader election.
This caused servers to crash in the presence of flaky DNS.

This patch changes the behavior to count a failed resolution as a denied vote.
So, if a server ends up with broken DNS for a few minutes, it will get into the
normal path of calling leader elections, counting them as failed, and trying
again later.

This ought to fix occasional server crashes we see on the YCSB cluster and
bolt80.

Change-Id: Ie4d623bc867e76972933530f14c16a261dd99630
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7870
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>