KUDU-882. Ensure CSD has the most important flags
This adds a few new flags to the CSD that weren't represented before.
It also makes the capitalization consistent and removes memory
cgroup auto-configuration. Adar and I are of the opinion that we
should rely on our memory limiting (which has nice service-preserving
behavior) vs Linux's (which is the OOM killer).

Change-Id: I2b8b21ff36f66d898022ccdc58881bacbf11b7c6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7878
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>