metrics: when filtering for a specific metric, don't print other entities
This improves the /metrics path handler such that if you are looking for
a specific metric, you only get entities dumped which contain that metric.
For example, '/metrics?metric=Write' is looking for a ts-wide metric, but
it still used to dump all of the tablet entities. Now it just dumps the
server entity which contains the Write RPC histogram metric.

Tested manually as well as a simple new unit test.

Change-Id: I52191e3b484139dc4fbf100e61bb0f515a93787e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7882
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>