KUDU-1114. Fix a crash when viewing /scans while a scanner is still initializing
This fixes an issue where we would try to access scanner->iter() without
checking for NULL first. Scanners are registered in the scanner manager before
they've initialized their iterator. In that case, the /scans web page would see
a NULL iterator and crash.

The fix is reasonable simple -- I added locking around Scanner::Init() and
added an IsInitialized(). This should be sufficient synchronization since we
never de-initialize a scanner after it is initialized.

I added a new thread to an existing unit test which grabs /scans and
/transactions in a loop while other threads do scans and transactions. This
reproduced the crash reliably but now passes.

Change-Id: I3ee89c21d5168940e6a481233a3121cbfa49fbc7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7885
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>