KUDU-86. Allow TS to retry registration RPC
This fixes an issue seen on bolt80 after enabling log fsyncing on the master.
With a lot of tablets, and fsyncing on, the master falls behind a little
bit at startup, since each registration ends up writing a lot of junk
to the log (we don't do a great job of coalescing these writes). So, a
few TS registrations end up timing out, which causes the TS to retry
registering.

Due to KUDU-86, the master was rejecting the retries, since it had in
fact processed the original registration.

This fixes the logic to allow a TS to re-register with the same sequence
number. I verified this allowed me to start the bolt80 master.

Change-Id: Id3075c79d658822825df46fbc70e1d3b30a81e27
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7891
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>