java: add a builder method to set the NIO executors
A previous commit removed the ability for the user to specify their
own SocketChannelFactory to the async client, since we've shaded
Netty. However, the most common use for this was to get our Netty
tasks running on the same thread pools as a user application.

This adds a builder method to AsyncKuduClientBuilder so that the
user can still specify the executors which will be passed to Netty.

In addition, I changed the default thread names to be more friendly
("kudu-nio-%d") and did a little code cleanup while I was at it.

Change-Id: Ib99dc8d7f553454666e938e70c6f32e7e1fad14d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7922
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>