KUDU-1129. Add Atomic CAS option to DeleteTablet()
Since the Master often has a stale view of tablet state, and we want it
to cull replicas that are no longer a part of the config, we need to
provide a way for it to specify the latest known config it is operating
with when deleting replicas. This adds an optional parameter to
DeleteTablet() that rejects the request when the caller specifies an
index of the committed config that is out of date.

Change-Id: I8920b45d70aabb17b5a9f26d03d7ae922f400d96
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7979
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>