KUDU-933. Master should delete replicas not in the latest raft config
This patch adds support for Master-initiated tablet deletion based on
the latest committed consensus config reported by tablet servers.

When we receive a heartbeat containing a newly committed raft config, we
immediately initiate a DeleteTablet request to any replica that was a
part of the old config but not part of the new one.

Additionally, when we receive a heartbeat from a replica that is not a
part of the committed config (from catalog_manager.cc):

We only tombstone replicas that include a committed raft config in their
report that has an opid_index strictly less than the latest reported
committed config, and (obviously) who are not members of the latest
config. This prevents us from spuriously deleting replicas that have
just been added to a pending config and are in the process of catching
up to the log entry where they were added to the config.

Change-Id: I27d420da2993684bc247496b11a27b5824d143d1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7983
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>